Added OSFile and updated FileSystem
Compacted OpenFile.java and SaveFile.java into one file OSFile.java. Updated FileSystem.java with a new GUI. Everything is connected with a working directory that the user will be able to change in the future. Added 'Add File', 'Retrieve File', and 'Remove File'. The first two work and the last one still needs to be implemented.